+++ src/main/java/xyz/qincai/manhunt/config/ConfigManager.java
@@ -98,6 +98,7 @@ public class ConfigManager {
 
         validateInt("hunters.respawnLimit", -1);
 
+        validateBoolean("chat.logToConsole", true);
         validateBoolean("scoreboard.enabled", true);
         validateBoolean("actionBar.enabled", true);
 
@@ -370,6 +371,10 @@ public class ConfigManager {
 
     public int getHunterRespawnLimit() {
         return config.getInt("hunters.respawnLimit", -1);
+    }
+
+    public boolean isChatLogToConsole() {
+        return config.getBoolean("chat.logToConsole", true);
     }
 
     public boolean isScoreboardEnabled() {