+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `chat.logToConsole` config option (default `true`) to log game chat
+  messages to the server console with format `[Manhunt Chat] [Global/Team] Play
+  r: message`
+
 ## [1.5.0] - 2026-07-10
 
 ### Added